+++ packages/blip/app/pages/patient/patientsettings.js
@@ -18,7 +18,6 @@ import PropTypes from 'prop-types';
 
 import React, { Component } from 'react';
 import _ from 'lodash';
-import sundial from 'sundial';
 import { utils } from 'tidepool-viz';
 import { translate, Trans } from 'react-i18next';